+++ src/components/InputNumber/constants.ts
@@ -1,6 +1,0 @@
-export const InputNumberMode = {
-  CURRENCY: 'currency',
-  DECIMAL: 'decimal'
-} as const;
-export type InputNumberMode =
-  (typeof InputNumberMode)[keyof typeof InputNumberMode];