+++ src/components/NumberInput/constants.ts
@@ -1,0 +1,6 @@
+export const NumberInputMode = {
+  CURRENCY: 'currency',
+  DECIMAL: 'decimal'
+} as const;
+export type NumberInputMode =
+  (typeof NumberInputMode)[keyof typeof NumberInputMode];