+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## v2.0.12
 
 - Removes the `CurrencyInput` component
-- Add the `InputNumber` component
+- Add the `NumberInput` component
 
 ## v2.0.11